fix: refactor variables parsing in domain parser
Since in previous commits we splitted parsing of names from parsing of variables, we can make more stricter assumptions on the input of the method. This allowed to simplify the code to a large extent.

This introduced some code duplication with the typed_list_names method; but the author's opinion is that it is a good-enough solution for the state of the codebase, where things can be refactored and changed rapidly. Moreover, note that, differently from the typed_list_names, typed_list_variables has to handle multiple types of variables, like the 'either' construct.