feat: changing domain parser behaviour on typed_list_name (backward compatible)
This commit changes the way typed_list_name rule is parsed. This does not change the parsing behaviour from the user perspective, but it is a non-functional change to improve the performances of the Lark parser.

With the previous version of the rule 'typed_list_name: NAME+ TYPE_SEP primitive_type (typed_list_name)', the Lark parser internals would perform a recursive call whenever a typed_list_name is matched. This might cause an arbitrarily long stack of calls whenever the parser got an input like:

```
element1 - t1 element2 - t1 element3 t1 ...
```

The above list of tokens is parsed as:

```
element1 - t1 <typed_list_name>
    element2 - t1 <typed_list_name>
        element3 - t1 <typed_list_name>
            ...
```

For large inputs, this will easily hit the recursion depth limit.

With the new approach, the entire list of tokens is parsed at once. This adds some complexity in the parsing function for the typed_list_name rule, but we have more control in how the parsing is performed; in particular, we can parse the tokens *iteratively* rather than *recursively*.

Finally, the NAME* pattern is appended at the end of the typed_list_name rule. This is because, according to the syntax specification, the last sublist of names might be non-typed.

The implementation exploits the newly added TypesIndex class which handles corner cases and syntax errors (e.g. duplicated entries in the list).