fix: remove type checking in Predicate.__call__
This commit removes a check on the types of the new terms passed as arguments to
Predicate.__call__. The reason is that the check is not "complete": for example,
the check returns false when a new term whose types are a subset of the types of
the term being replaced, while the correct result is true.

More fundamentally, the problem is that the class Predicate, in its current
state, does not have information on the type hierarchy of the domain the
predicate belongs to. Having this check here adds more problems than benefits.
The final type check will be performed when instantiating the Domain class while
leaving the flexibility to the user to instantiate temporary predicate objects,
even with terms whose type is incompatible with the target type hierarchy of the
PDDL domain.

Therefore, rather than having a half-working feature, it is better to remove it
altogether. The idea is still relevant, but it may require a distinction between
'PredicateDefinition' (which has a pointer to the domain or the type hierarchy)
and 'Predicate' (the actual instance of the predicate definition, with only
constants).